Share a zeroconf instance between homekit bridges
With camera support it is likely users will have
2-3 bridges.  We want to avoid having to read all
the zeroconf traffic on the network multiple times
only to make a single homekit bridge announcement.

Additionally since zeroconf keeps a cache, for network
the size of a /20 with a lot of zeroconf traffic this
worked out to ~14MiB per zeroconf instance.